# By Carlos C (56) and others
# Via Carlos C (17) and others
* 'master' of https://github.com/software-engineering-amsterdam/many-ql: (198 commits)
  Finished expressions in AST
  refactored evaluator + support unary op
  Add true and false as reserved words
  Added first draft of a type checker.
  type check progress
  Fix typechecker double evaluation of identifiers
  Code Grooming
  Improve naming consistency
  Fix regression in type checker tests
  Detect repeated question labels
  "fix" GUI stuff, evaluator turned off
  Validate expressions
  Extend Execute visitor into a Typechecker visitor
  Fixed grammar
  - Added a little bit of a solution for expression evaluation.
  typecheck init
  - Fixed some bugs with adding the references to the ReferenceMap
  Added some properties to the AST.
  - Added a specific ReferenceMap with custom method to find Reference to a specific type. - Separated typechecking and build up of reference map. - Expressions now have optional operators. - Altered grammar for operators (they are now Tokens).
  Improve typechecker and cover with tests
  ...

+++ KrasiMilan/QL/src/main/java/nl/uva/softwcons/ast/expression/literal/IntegerLiteral.java
@@ -1,14 +1,16 @@
 package nl.uva.softwcons.ast.expression.literal;
 
+import java.math.BigInteger;
+
 import nl.uva.softwcons.ast.expression.Expression;
 import nl.uva.softwcons.ast.expression.ExpressionVisitor;
 
 public class IntegerLiteral extends Expression {
 
-    private int value;
+    private BigInteger value;
 
     public IntegerLiteral(int value) {
-        this.value = value;
+        this.value = BigInteger.valueOf(value);
     }
 
     @Override
@@ -16,7 +18,7 @@ public class IntegerLiteral extends Expression {
         return visitor.visit(this);
     }
 
-    public int getValue() {
+    public BigInteger getValue() {
         return value;
     }
